requirements.txt
pandas~=1.4.2
transformers~=4.28.0
rdflib~=7.0.0
sentence-transformers~=2.2.2
ipywidgets~=8.1.2
scikit-learn~=1.4.0
Jinja2==3.1.3
fastapi==0.110.2
uvicorn==0.29.0
setuptools==69.5.1
SQLAlchemy==2.0.29
python-multipart==0.0.9
torch~=2.3.0
numpy~=1.26.4
starlette~=0.37.2
requests~=2.32.3
python-dotenv~=1.0.1
matplotlib~=3.9.2